Solution for syncing spreadsheet sizes automatically. Added spreadsheetSyncSizesWrappers for Element_Page. Added hideMultiline to Page, which affects all Labels inside it. Changed Label's hideMultiline parameter to default to it's parentPage.
Added a 'hideMultiline' option to Label.
Enabled this for all Labels in MainGrid in Spreadsheet.
Put all _sync functions in _syncSizes() and bound them to '<Configure>' of MainGrid when df has been loaded.

Made spreadsheet prettier.

Changed default font to Consolas which is fixed-length and has no kerning. Aka monospace.